pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "turbollama"
version = "1.0.0"
description = "The Ultimate llama.cpp Wrapper with Integrated GUI - Faster than Ollama"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "AI Joe", email = "ai.joe.git@gmail.com"}
]
maintainers = [
    {name = "AI Joe", email = "ai.joe.git@gmail.com"}
]
keywords = [
    "llama", "llama.cpp", "ollama", "ai", "llm", "inference", 
    "gpu", "gradio", "openai", "chatbot", "machine-learning"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "gradio>=4.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "aiofiles>=23.0.0",
    "python-multipart>=0.0.6",
    "llama-cpp-python>=0.2.20",
    "torch>=2.0.0",
    "huggingface-hub>=0.17.0",
    "transformers>=4.35.0",
    "psutil>=5.9.0",
    "GPUtil>=1.4.0",
    "py-cpuinfo>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pre-commit>=3.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]
intel = [
    "intel-extension-for-pytorch",
]
all = [
    "turbollama[dev,intel]",
]

[project.urls]
Homepage = "https://github.com/ai-joe-git/TurboLlama"
Documentation = "https://turbollama.readthedocs.io"
Repository = "https://github.com/ai-joe-git/TurboLlama.git"
"Bug Tracker" = "https://github.com/ai-joe-git/TurboLlama/issues"
Changelog = "https://github.com/ai-joe-git/TurboLlama/blob/main/CHANGELOG.md"

[project.scripts]
turbollama = "turbollama.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["turbollama*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
